+++ extensions/search-npm/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Search npm Changelog
 
+## [Update] - 2024-12-13
+
+- Added a Git URL parser for the repository URL in the open repository action
+
+## [Updates] - 2024-10-07
+
+- Hide history list when `historyCount` is zero
+- Bump all dependencies to the latest
+
 ## [Refactoring] - 2024-02-03
 
 - Bump dependecies